improve ui

+++ web/src/components/admin/settings/AdminAgentsTab.vue
@@ -48,7 +48,7 @@
             <TextField v-model="selectedAgent.token" :placeholder="$t('admin.settings.agents.token')" disabled />
           </InputField>
 
-          <InputField :label="$t('admin.settings.agents.backend')">
+          <InputField :label="$t('admin.settings.agents.backend')" docs-url="docs/next/administration/backends/docker">
             <TextField v-model="selectedAgent.backend" disabled />
           </InputField>
 
@@ -56,8 +56,12 @@
             <TextField v-model="selectedAgent.platform" disabled />
           </InputField>
 
-          <InputField :label="$t('admin.settings.agents.capacity')">
-            <TextField v-model="selectedAgent.capacity" disabled />
+          <InputField
+            :label="$t('admin.settings.agents.capacity')"
+            docs-url="docs/next/administration/agent-config#woodpecker_max_procs"
+          >
+            <span class="text-color-alt">The max amount of parallel pipelines executed by this agent.</span>
+            <TextField :model-value="selectedAgent.capacity?.toString()" disabled />
           </InputField>
 
           <InputField :label="$t('admin.settings.agents.last_contact')">
@@ -89,6 +93,8 @@ import { useI18n } from 'vue-i18n';
 
 import Button from '~/components/atomic/Button.vue';
 import ListItem from '~/components/atomic/ListItem.vue';
+import InputField from '~/components/form/InputField.vue';
+import TextField from '~/components/form/TextField.vue';
 import Panel from '~/components/layout/Panel.vue';
 import useApiClient from '~/compositions/useApiClient';
 import { useAsyncAction } from '~/compositions/useAsyncAction';
@@ -127,6 +133,11 @@ const { doSubmit: saveAgent, isLoading: isSaving } = useAsyncAction(async () => 
 });
 
 const { doSubmit: deleteAgent, isLoading: isDeleting } = useAsyncAction(async (_agent: Agent) => {
+  // eslint-disable-next-line no-restricted-globals, no-alert
+  if (!confirm(i18n.t('admin.settings.agents.delete_confirm'))) {
+    return;
+  }
+
   await apiClient.deleteAgent(_agent);
   notifications.notify({ title: i18n.t('admin.settings.agents.deleted'), type: 'success' });
   await loadAgents();
